Move contributors actions from projectRepository to the projectContributors repository
